python/DS-Project/new/requirements.txt
numpy
Pillow
Flask
matplotlib
memory-profiler